Reports

+++ my expenses/Expenses.cs
@@ -2,7 +2,7 @@
 {
     public class Expenses
     {
-        public int Amount { get; set; }
+        public Int64 Amount { get; set; }
         public string cards { get; set; }
         public string Grouping { get; set; }
         public string Description { get; set; }
